Get data from firestore instead of dummydatas
