+++ src/components/QuestionList/QuestionList.tsx
@@ -1,17 +1,16 @@
-import { FC, useEffect, useState } from 'react';
+import { FC } from 'react';
 import { useHistory, useLocation } from 'react-router';
 import styled from 'styled-components';
 import { Box, List, Typography } from '@material-ui/core';
 import { H3, TRUNCATE_ONE, LIGHT_GRAY_1 } from 'utils/themes';
-import firebase from 'firebase';
-import { Topic, SubTopic, Question, QuestionFB, AnswerContent, QuestionContent } from 'utils/types';
-import { TimestampToDate } from 'utils/functions';
+import { Topic, SubTopic, Question } from 'utils/types';
 import { Hover } from 'components/Contents';
 import { QuestionListElement } from './QuestionListElement';
 
 interface QuestionListHeaderProp {
   topic: Topic;
   subTopic: SubTopic;
+  questionList: Question[];
   isListShown: boolean;
   questionId?: number;
   questionId2?: number;
@@ -25,6 +24,7 @@ interface QuestionListHeaderProp {
 export const QuestionList: FC<QuestionListHeaderProp> = ({
   topic,
   subTopic,
+  questionList,
   isListShown,
   questionId,
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -37,44 +37,6 @@ export const QuestionList: FC<QuestionListHeaderProp> = ({
 }) => {
   const history = useHistory();
   const location = useLocation();
-
-  const [questionIdList, setQuestionIdList] = useState<number[]>(subTopic.questionList as number[]);
-  const [questionList, setQuestionList] = useState<Question[]>();
-
-  // setQuestionIdList(subTopic.questionList as number[]);
-  useEffect(() => {
-    setQuestionIdList(subTopic.questionList as number[]);
-  }, [subTopic]);
-
-  useEffect(() => {
-    if (questionIdList !== undefined) {
-      firebase
-        .firestore()
-        .collection('questions')
-        .get()
-        .then((doc) => {
-          const questionListCustom = [] as Question[];
-          doc.docs.filter((item) => {
-            const { question, answers, ...rest } = item.data() as QuestionFB;
-            const questionContent = { ...question, time: TimestampToDate(question.time) } as QuestionContent;
-            const answerContents = answers.map(
-              (item) => ({ ...item, time: TimestampToDate(item.time) } as AnswerContent),
-            );
-
-            const finalQuestion = { question: questionContent, answers: answerContents, ...rest } as Question;
-
-            if (questionIdList.includes(finalQuestion.questionId)) {
-              questionListCustom.push(finalQuestion);
-            }
-
-            return finalQuestion;
-          });
-          questionListCustom.sort((a, b) => b.questionId - a.questionId);
-          setQuestionList(questionListCustom);
-        })
-        .catch();
-    }
-  }, [questionIdList]);
 
   const onClickItem = (item: Question) => {
     const path = location.pathname;
