+++ src/pages/NormalQuestion.tsx
@@ -3,11 +3,13 @@ import { FC, useEffect, useState } from 'react';
 import { useHistory } from 'react-router';
 import styled from 'styled-components';
 import { Backdrop, Box, Fade, Paper } from '@material-ui/core';
-import { PageType, Topic, SubTopic } from 'utils/types';
+import firebase from 'firebase';
+import { PageType, Topic, SubTopic, Question, QuestionFB, QuestionContent, AnswerContent } from 'utils/types';
+import { TimestampToDate } from 'utils/functions';
 import { QuestionList } from 'components/QuestionList/QuestionList';
 import { Hover, Contents, NotSelected } from 'components/Contents';
 import { useTopicList } from 'hooks/useTopicList';
-import { dummyQuestions } from '../utils/dummyDatas';
+import { useQuestionList } from 'apis/Question/useQuestionList';
 
 interface NormalQuestionProp {
   pageType: PageType;
@@ -42,10 +44,16 @@ export const NormalQuestion: FC<NormalQuestionProp> = ({
 }) => {
   const [topicInfo, setTopicInfo] = useState<Topic>();
   const [subTopicInfo, setSubTopicInfo] = useState<SubTopic>();
+  const [questionIdList, setQuestionIdList] = useState<number[]>();
+  const [questionList, setQuestionList] = useState<Question[]>();
+  const [question1, setQuestion1] = useState<Question>();
+  const [question2, setQuestion2] = useState<Question>();
 
   const { topicList } = useTopicList();
 
   const history = useHistory();
+
+  /** Close questions */
 
   const onCloseLeftContent = () => {
     if (questionId2 !== undefined) {
@@ -79,13 +87,64 @@ export const NormalQuestion: FC<NormalQuestionProp> = ({
     }
   }, [topicInfo, topicId, subTopicId]);
 
+  /** Get Question list */
+
+  useEffect(() => {
+    if (subTopicInfo !== undefined) {
+      setQuestionIdList(subTopicInfo.questionList as number[]);
+    }
+  }, [subTopicInfo]);
+
+  useEffect(() => {
+    if (questionIdList !== undefined) {
+      firebase
+        .firestore()
+        .collection('questions')
+        .get()
+        .then((doc) => {
+          const questionListCustom = [] as Question[];
+          doc.docs.filter((item) => {
+            const { question, answers, ...rest } = item.data() as QuestionFB;
+            const questionContent = { ...question, time: TimestampToDate(question.time) } as QuestionContent;
+            const answerContents = answers.map(
+              (item) => ({ ...item, time: TimestampToDate(item.time) } as AnswerContent),
+            );
+
+            const finalQuestion = { question: questionContent, answers: answerContents, ...rest } as Question;
+
+            if (questionIdList.includes(finalQuestion.questionId)) {
+              questionListCustom.push(finalQuestion);
+            }
+
+            return finalQuestion;
+          });
+          questionListCustom.sort((a, b) => b.questionId - a.questionId);
+          setQuestionList(questionListCustom);
+        })
+        .catch();
+    }
+  }, [questionIdList]);
+
+  useEffect(() => {
+    if (questionList !== undefined) {
+      setQuestion1(questionList.find((question) => question.questionId === questionId));
+    }
+  }, [questionList, questionId]);
+
+  useEffect(() => {
+    if (questionList !== undefined) {
+      setQuestion2(questionList.find((question) => question.questionId === questionId2));
+    }
+  }, [questionList, questionId2]);
+
   return (
     <QuestionsContainer>
       <QuestionDetails>
-        {topicInfo !== undefined && subTopicInfo !== undefined && (
+        {topicInfo !== undefined && subTopicInfo !== undefined && questionList !== undefined && (
           <QuestionList
             topic={topicInfo}
             subTopic={subTopicInfo}
+            questionList={questionList}
             questionId={questionId}
             questionId2={questionId2}
             isListShown={isListShown}
@@ -101,12 +160,16 @@ export const NormalQuestion: FC<NormalQuestionProp> = ({
           {questionId === undefined && <NotSelected />}
           {questionId !== undefined && (
             <QBox>
-              <Contents question={dummyQuestions[questionId - 1]} closeThisContent={onCloseLeftContent} />
+              {questionList !== undefined && question1 !== undefined && (
+                <Contents question={question1} closeThisContent={onCloseLeftContent} />
+              )}
             </QBox>
           )}
           {questionId2 !== undefined && (
             <QBox>
-              <Contents question={dummyQuestions[questionId2 - 1]} closeThisContent={onCloseRightContent} />
+              {questionList !== undefined && question2 !== undefined && (
+                <Contents question={question2} closeThisContent={onCloseRightContent} />
+              )}
             </QBox>
           )}
         </QQBox>
